+++ client/package.json
@@ -8,6 +8,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "antd": "^5.8.2",
     "graphql": "^16.7.1",
     "jwt-decode": "^3.1.2",
     "react": "^18.2.0",
