+++ client/src/components/ProfileSwitcher.jsx
@@ -1,0 +1,34 @@
+import React, { useState } from 'react';
+import Avatar from '@material-ui/core/Avatar';
+import Menu from '@material-ui/core/Menu';
+import MenuItem from '@material-ui/core/MenuItem';
+
+const ProfileSwitcher = () => {
+  const [anchorEl, setAnchorEl] = useState(null);
+
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget);
+  };
+
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
+
+  return (
+    <div>
+      <Avatar onClick={handleClick}>U</Avatar>
+      <Menu
+        id="simple-menu"
+        anchorEl={anchorEl}
+        keepMounted
+        open={Boolean(anchorEl)}
+        onClose={handleClose}
+      >
+        <MenuItem onClick={handleClose}>Gardener Profile</MenuItem>
+        <MenuItem onClick={handleClose}>Homeowner Profile</MenuItem>
+      </Menu>
+    </div>
+  );
+}
+
+export default ProfileSwitcher;
