+++ client/src/pages/ProfilePage.jsx
@@ -1,57 +1,71 @@
-import React, { useState } from "react";
-import { gql, useQuery, useMutation } from "@apollo/client";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import PlotCard from "../components/PlotCard";
-import AddPlot from "../components/AddPlot";
-import { GET_PLOTS, ADD_PLOT, EDIT_PLOT, DELETE_PLOT } from "../utils/mutations";
-import { set } from "mongoose";
-import { faPlus } from "@fortawesome/free-solid-svg-icons";
-import { useAuth } from "../utils/AuthContext";
+import React, { useState, useContext, useEffect } from "react";
+import { gql, useQuery } from "@apollo/client";
+import { AuthContext } from "../utils/AuthContext";
+import { Menu, Dropdown, Button, Avatar } from 'antd';
+import GardenerProfileComponent from "../components/GardenerProfileComponent";
+import HomeownerProfileComponent from "../components/HomeownerProfileComponent";
+import { GET_CURRENT_USER } from "../utils/mutations";
 
-const plusIcon = <FontAwesomeIcon icon={faPlus} />;
+const MasterProfilePage = () => {
+  const { currentUser } = useContext(AuthContext);
+  const { loading, error, data } = useQuery(GET_CURRENT_USER);
 
+  const [profile, setProfile] = useState(null); // Initialize with null
 
-const ProfilePage = () => {
-  const { loading, error, data } = useQuery(GET_PLOTS);
-  const [showAddPlotModal, setShowAddPlotModal] = useState(false);
-  const [selectedPlot, setSelectedPlot] = useState(null);
+  useEffect(() => {
+    if (data && data.getCurrentUser) {
+      const { isGardener, isHomeowner } = data.getCurrentUser;
+      // If user is a gardener, set profile to 'gardener' by default
+      // If user is not a gardener but a homeowner, set profile to 'homeowner'
+      // If user is neither a gardener nor a homeowner, set profile to null
+      if (isGardener) {
+        setProfile('gardener');
+      } else if (isHomeowner) {
+        setProfile('homeowner');
+      } else {
+        setProfile(null);
+      }
+    }
+  }, [data]);
 
-  const toggleAddPlotModal = (plot) => {
-    setSelectedPlot(plot);
-    setShowAddPlotModal(!showAddPlotModal);
+  const toggleGardenerProfile = () => {
+    setProfile(profile === 'gardener' ? null : 'gardener');
+  };
+
+  const toggleHomeownerProfile = () => {
+    setProfile(profile === 'homeowner' ? null : 'homeowner');
   };
 
   if (loading) return <p>Loading...</p>;
-  if (error) {
-    console.log("Error message:", error.message);
-    console.log("Full error object:", error);
-    return <p>Error :(</p>;
-  }
+  if (error) return <p>Error :(</p>;
+
+  const userFirstName = data.getCurrentUser.firstName;
+  const userLastName = data.getCurrentUser.lastName;
+  const userInitials = `${userFirstName.charAt(0)}${userLastName.charAt(0)}`.toUpperCase();
+
+  const menu = (
+    <Menu>
+      <Menu.Item key="1" onClick={toggleGardenerProfile}>
+        Gardener Profile
+      </Menu.Item>
+      <Menu.Item key="2" onClick={toggleHomeownerProfile}>
+        Homeowner Profile
+      </Menu.Item>
+    </Menu>
+  );
 
   return (
     <div>
-      <h1 className="text-4xl text-yard-red text-center my-8">Your Plots</h1>
-      <div className="plot-list flex space-x-4 flex-wrap justify-center">
-      {data.getPlots.map((plot) => (
-        <PlotCard key={plot.id} plot={plot} onClick={() => toggleAddPlotModal(plot)} />
-      ))}
-      </div>
-      {showAddPlotModal && (
-        <div className="modal">
-          <div className="modal-content">
-            <span className="close-button" onClick={() => toggleAddPlotModal(null)}>
-              X
-            </span>
-            <AddPlot plot={selectedPlot} />
-          </div>
-        </div>
-      )}
-      <div className="add-plot-button flex justify-center flex-col items-center">
-      <p className="text-yard-orange">Add New Plot</p>
-      <button className="spin-button text-yard-orange text-6xl p-0 m-0 " onClick={() => toggleAddPlotModal(null)}>{plusIcon}</button>
-      </div>
+      <h1>Your Profile</h1>
+      <Dropdown overlay={menu}>
+        <span>
+          <Avatar style={{ cursor: "pointer" }} size="large">{userInitials}</Avatar> {/* Replace 'U' with user's initials */}
+        </span>
+      </Dropdown>
+      {profile === 'gardener' && <GardenerProfileComponent />}
+      {profile === 'homeowner' && <HomeownerProfileComponent />}
     </div>
   );
 };
 
-export default ProfilePage;
+export default MasterProfilePage;
